Renamespace identifier validator
